lexer quote
