+++ Makefile
@@ -4,7 +4,7 @@ LIBS    = libft/libft.a
 INCLUDE = -I./include -I./libft/include -I $(brew --prefix readline)/include
 NAME    = minishell
 SRCDIR  = src
-SRCS    = src/lexer/lexer_helper.c src/lexer/lexer.c src/utils/error.c src/heredoc/heredoc.c src/parser/ast.c src/parser/parser.c src/main.c src/exec/file_access_util.c src/exec/exec.c src/exec/list_to_array.c src/exec/execute_cmd.c src/builtin/env_lst_operations.c src/builtin/decompress.c src/builtin/mini_echo.c src/builtin/mini_cd.c src/builtin/mini_export.c src/builtin/mini_handle_command.c src/builtin/mini_unset.c src/builtin/mini_env.c src/builtin/mini_pwd.c src/builtin/mini_exit.c
+SRCS    = src/lexer/lexer_helper.c src/lexer/lexer.c src/lexer/token.c src/utils/error.c src/heredoc/heredoc.c src/parser/ast.c src/parser/parser.c src/main.c src/exec/file_access_util.c src/exec/exec.c src/exec/list_to_array.c src/exec/execute_cmd.c src/builtin/env_lst_operations.c src/builtin/decompress.c src/builtin/mini_echo.c src/builtin/mini_cd.c src/builtin/mini_export.c src/builtin/mini_handle_command.c src/builtin/mini_unset.c src/builtin/mini_env.c src/builtin/mini_pwd.c src/builtin/mini_exit.c
 OBJDIR  = obj
 OBJS    = $(subst $(SRCDIR), $(OBJDIR), $(SRCS:.c=.o))
 DEPENDS = $(OBJS:.o=.d)
