+++ include/lexer.h
@@ -6,12 +6,15 @@
 /*   By: tsishika <tsishika@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/24 12:19:08 by tkuramot          #+#    #+#             */
-/*   Updated: 2023/09/19 22:06:06 by tkuramot         ###   ########.fr       */
+/*   Updated: 2023/09/19 23:54:39 by tkuramot         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef LEXER_H
 # define LEXER_H
+
+# define SINGLE_QUOTE '\''
+# define DOUBLE_QUOTE '\"'
 
 # include "libft.h"
 # include <stdbool.h>
@@ -44,6 +47,7 @@ bool	is_metacharacter(char c);
 bool	start_with(char *s, const char *prefix);
 bool	is_word(char c);
 void	consume_blank(char **line);
+t_token	*token_init(char *word, t_token_type type);
 t_token	*token_copy(t_token *token);
 void	token_lst_free(t_token *lst);
