+++ src/lexer/lexer.c
@@ -6,32 +6,13 @@
 /*   By: tsishika <tsishika@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/24 12:20:26 by tkuramot          #+#    #+#             */
-/*   Updated: 2023/09/18 23:24:43 by tkuramot         ###   ########.fr       */
+/*   Updated: 2023/09/20 00:00:19 by tkuramot         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "lexer.h"
 #include "utils.h"
 #include <stdio.h>
-
-static t_token	*token_init(char *word, t_token_type type)
-{
-	t_token	*token;
-
-	if (!word)
-		fatal_error("malloc");
-	token = ft_calloc(1, sizeof (t_token));
-	if (!token)
-		fatal_error("malloc");
-	token->word = word;
-	token->type = type;
-	return (token);
-}
-
-t_token	*token_copy(t_token *token)
-{
-	return (token_init(token->word, token->type));
-}
 
 static t_token	*extract_word(char **line)
 {
@@ -42,7 +23,21 @@ static t_token	*extract_word(char **line)
 	tmp = *line;
 	i = 0;
 	while (tmp[i] && is_word(tmp[i]))
+	{
+		if (tmp[i] == SINGLE_QUOTE)
+		{
+			while (tmp[++i] && tmp[i] != SINGLE_QUOTE)
+				i++;
+		}
+		else if (tmp[i] == DOUBLE_QUOTE)
+		{
+			while (tmp[++i] && tmp[i] != DOUBLE_QUOTE)
+			{
+				i++;
+			}
+		}
 		i++;
+	}
 	token = token_init(ft_substr(tmp, 0, i), TK_WORD);
 	*line += i;
 	return (token);
@@ -94,18 +89,4 @@ t_token	*tokenize(char *line)
 		}
 	}
 	return (head.next);
-}
-
-// リークが残って気になったので勝手に作りました。
-void	token_lst_free(t_token *lst)
-{
-	t_token	*buf;
-
-	while (lst)
-	{
-		buf = lst;
-		free(lst->word);
-		lst = lst->next;
-		free(buf);
-	}
 }
